+++ necsim/plugins/species/src/lib.rs
@@ -3,10 +3,9 @@
 #[macro_use]
 extern crate log;
 
-mod arrow;
-mod feather;
 mod identity;
-mod individuals;
+mod individual;
+mod location;
 mod state;
 
 use identity::SpeciesIdentity;
@@ -14,7 +13,7 @@ use state::LastEventState;
 
 // Register the reporter plugins
 necsim_plugins_core::export_plugin!(
-    IndividualSpeciesSQL => individuals::IndividualLocationSpeciesReporter,
-    IndividualSpeciesArrow => arrow::IndividualLocationSpeciesReporter,
-    IndividualSpeciesFeather => feather::LocationGroupedSpeciesReporter,
+    IndividualSpeciesSQLite => individual::sqlite::IndividualSpeciesSQLiteReporter,
+    IndividualSpeciesFeather => individual::feather::IndividualSpeciesFeatherReporter,
+    LocationSpeciesFeather => location::feather::LocationSpeciesFeatherReporter,
 );
