+++ necsim/plugins/species/src/location/mod.rs
@@ -1,0 +1,1 @@
+pub mod feather;
